refactor(diffshub): rename cn utility module
Rename lib/utils.ts to lib/cn.ts and update DiffsHub component imports plus the local components config alias.
