+++ apps/diffshub/components/StatusRow.tsx
@@ -1,6 +1,6 @@
 import type { ComponentType, ReactNode } from 'react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/cn';
 
 interface StatusRowProps {
   icon: ComponentType<{ className?: string }>;
